test: add retry package for retriable tests
The current retry framework in testutil/testprc.WaitForResult uses
callbacks to execute a function until it succeeds. It accepts a boolean
parameter as ok/nok and captures the error if any. At the end the test
is failed with Fatal and the last error.

This skews the line numbers in the error messages and makes the test
functions more complex than they need to be since both the bool value
and the error can indicate whether the operation succeeded or not.
Therefore, one of them is redundant.

The new retry framework uses a simple for-loop without callbacks to
retain the correct callstack since the Go testing lib does not offer an
option to specify the callstack depth for correct line numbers.